move AuthNavbar, AuthInput to folder atoms
