+++ src/pages/Auth/Login.jsx
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom';
 import { toast, ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import Button from '../../components/Button';
-import AuthInput from '../../components/molecules/AuthInput';
+import AuthInput from '../../components/atoms/AuthInput';
 import { setLogin } from '../../services/auth-api';
 
 function Login() {
